Тестовый комит

+++ LongWorkMP/InterworkingModule/Task.cs
@@ -2,16 +2,16 @@
 {
     public struct Task
     {
-        public Task(int rangeStart, int rangeEnd, string md5Sum)
+        public Task(long rangeStart, long rangeEnd, string md5Sum)
         {
             RangeStart = rangeStart;
             RangeEnd = rangeEnd;
             Md5Sum = md5Sum;
         }
 
-        public int RangeStart { get; set; }
+        public long RangeStart { get; set; }
 
-        public int RangeEnd { get; set; }
+        public long RangeEnd { get; set; }
 
         public string Md5Sum { get; set; }
     }
